Problemas 8-11 OmegaUp

+++ EjerciciosOmegaUp/SeleccionDeEstudiantes.java
@@ -1,11 +1,11 @@
 import java.util.Scanner;
-
+//FIXME
 public class SeleccionDeEstudiantes {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextInt(), b = sc.nextInt();
+        long a = sc.nextInt(), b = sc.nextInt();
         sc.close();
-        String str = Integer.toString(a * b);
+        String str = Long.toString(a * b);
         System.out.println(str.charAt(str.length()-1));
     }
 }
